backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
scikit-surprise==1.1.4
sentence-transformers==3.0.1
numpy==1.26.4
supabase==2.5.0
python-dotenv==1.0.1
scikit-learn==1.5.0
httpx==0.27.0